added input DTO's for multiple models.

+++ src/main/java/nl/jordy/petplacer/models/Donation.java
@@ -6,6 +6,8 @@ import lombok.AccessLevel;
 import lombok.Data;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Entity
 @Data
 @Table(name = "donations")
@@ -22,7 +24,7 @@ public class Donation {
     @ManyToOne
     private Shelter recievingShelter;
 
-    private double donationAmount;
+    private BigDecimal donationAmount;
     private String donationMessage;
 
 }
